Update print analytics component
