+++ mtp_common/assets-src/mtp_common/components/print-analytics/index.js
@@ -1,0 +1,18 @@
+// Tracks page printing
+'use strict';
+
+import {Analytics} from '../analytics';
+
+export var PrintAnalytics = {
+  init: function () {
+    if (window.print === undefined) {
+      return;
+    }
+    window.print = (function (printfn) {
+      return function () {
+        Analytics.send('event', 'print');
+        printfn();
+      };
+    }(window.print));
+  }
+};
